Nuevo de clase con BD

+++ marketplace/src/main/java/com/uade/tpo/marketplace/entity/dto/AuthResponseDto.java
@@ -3,7 +3,7 @@ package com.uade.tpo.marketplace.entity.dto;
 public record AuthResponseDto(
     String accessToken,
     String tokenType,
-    Integer userId,
+    Long userId,
     String username,
     String role
 ) {}
