+++ marketplace/src/main/java/com/uade/tpo/marketplace/entity/dto/DigitalKeyDto.java
@@ -3,8 +3,8 @@ package com.uade.tpo.marketplace.entity.dto;
 import java.time.Instant;
 
 public record DigitalKeyDto(
-    Integer id,
-    Integer productId,
+    Long id,
+    Long productId,
     String keyMask,
     String status,
     Instant createdAt,
