+++ marketplace/src/main/java/com/uade/tpo/marketplace/entity/dto/OrderItemCreateDto.java
@@ -3,6 +3,6 @@ package com.uade.tpo.marketplace.entity.dto;
 import jakarta.validation.constraints.*;
 
 public record OrderItemCreateDto(
-    @NotNull Integer productId,
+    @NotNull Long productId,
     @NotNull @Min(1) Integer quantity
 ) {}
