+++ marketplace/src/main/java/com/uade/tpo/marketplace/entity/dto/OrderItemDto.java
@@ -3,7 +3,7 @@ package com.uade.tpo.marketplace.entity.dto;
 import java.math.BigDecimal;
 
 public record OrderItemDto(
-    Integer productId,
+    Long productId,
     String productTitle,
     BigDecimal unitPrice,
     Integer quantity,
