+++ marketplace/src/main/java/com/uade/tpo/marketplace/entity/dto/ReviewCreateDto.java
@@ -4,7 +4,7 @@ package com.uade.tpo.marketplace.entity.dto;
 import jakarta.validation.constraints.*;
 
 public record ReviewCreateDto(
-    @NotNull Integer productId,
+    @NotNull Long productId,
     @NotNull Integer rating,
     String title,
     String comment
