+++ marketplace/src/main/java/com/uade/tpo/marketplace/repository/interfaces/IOrderItemRepository.java
@@ -5,11 +5,12 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.uade.tpo.marketplace.entity.basic.DigitalKey;
 import com.uade.tpo.marketplace.entity.basic.OrderItem;
 
 
-public interface IOrderItemRepository extends JpaRepository<OrderItem, Integer> {
-    Optional<OrderItem> getOrderItemById(Integer id);
-    List<OrderItem> getOrderItemsByOrderId(Integer orderId);
+public interface IOrderItemRepository extends JpaRepository<OrderItem, Long>{
+    Optional<OrderItem> getOrderItemById(Long id);
+    List<OrderItem> getOrderItemsByOrderId(Long orderId);
     OrderItem creaOrderItem(OrderItem item);
 }
