+++ marketplace/src/main/java/com/uade/tpo/marketplace/repository/interfaces/IOrderRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.uade.tpo.marketplace.entity.basic.Order;
 
 
-public interface IOrderRepository extends JpaRepository<Order, Integer> {
-    Optional<Order> getOrderById(Integer id);
-    List<Order> getOrdersByBuyerId(Integer buyerId);
+public interface IOrderRepository extends JpaRepository<Order, Long>{
+    Optional<Order> getOrderById(Long id);
+    List<Order> getOrdersByBuyerId(Long buyerId);
     List<Order> getOrders();
     Order creatOrder(Order order);
-    void deleteOrderById(Integer id);
+    void deleteOrderById(Long id);
 }
