+++ marketplace/src/main/java/com/uade/tpo/marketplace/repository/interfaces/IProductRepository.java
@@ -5,14 +5,15 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.uade.tpo.marketplace.entity.basic.OrderItem;
 import com.uade.tpo.marketplace.entity.basic.Product;
 
 
-public interface IProductRepository extends JpaRepository<Product, Integer> {
+public interface IProductRepository extends JpaRepository<Product, Long>{
     List<Product> getProducts();
     List<Product> getActiveProducts();
-    List<Product> getProductsByCategoryId(Integer categoryId);
-    Optional<Product> getProductById(Integer id);
+    List<Product> getProductsByCategoryId(Long categoryId);
+    Optional<Product> getProductById(Long id);
     Product createProduct(Product product);
-    void deleteProductById(Integer id);
+    void deleteProductById(Long id);
 }
