+++ marketplace/src/main/java/com/uade/tpo/marketplace/repository/interfaces/IUserRepository.java
@@ -5,14 +5,15 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.uade.tpo.marketplace.entity.basic.OrderItem;
 import com.uade.tpo.marketplace.entity.basic.User;
 
 
-public interface IUserRepository extends JpaRepository<User, Integer> {
-    Optional<User> getUserById(Integer id);
+public interface IUserRepository extends JpaRepository<User, Long>{
+    Optional<User> getUserById(Long id);
     Optional<User> getUserByUsername(String username);
     Optional<User> getUserByEmail(String email);
     User creatUser(User user);
     List<User> getUsers();
-    void deleteUserById(Integer id);
+    void deleteUserById(Long id);
 }
